+++ src/Mapper/Category.php
@@ -10,6 +10,7 @@ use Doctrine\ORM\ORMException;
 use Doctrine\ORM\Tools\Pagination\Paginator;
 use jtl\Connector\Core\Exception\LanguageException;
 use jtl\Connector\Core\Utilities\Language as LanguageUtil;
+use \jtl\Connector\Shopware\Utilities\Locale as LocaleUtil;
 use jtl\Connector\Model\Category as JtlCategory;
 use jtl\Connector\Model\Identity;
 use jtl\Connector\Shopware\Model\CategoryAttr;
@@ -212,7 +213,7 @@ class Category extends DataMapper
         if (is_null($swCategory)) {
             $name = null;
             foreach ($jtlCategory->getI18ns() as $i18n) {
-                if (LanguageUtil::map(null, null, $i18n->getLanguageISO()) === Shopware()->Shop()->getLocale()->getLocale()) {
+                if (ShopUtil::isShopwareDefaultLanguage($i18n->getLanguageISO())) {
                     $name = $i18n->getName();
                     break;
                 }
@@ -247,7 +248,7 @@ class Category extends DataMapper
         // I18n
         $exists = false;
         foreach ($jtlCategory->getI18ns() as $i18n) {
-            if (LanguageUtil::map(null, null, $i18n->getLanguageISO()) === Shopware()->Shop()->getLocale()->getLocale()) {
+            if (ShopUtil::isShopwareDefaultLanguage($i18n->getLanguageISO())) {
                 $exists = true;
 
                 $swCategory->setName($i18n->getName());
@@ -277,7 +278,9 @@ class Category extends DataMapper
     protected function prepareAttributeAssociatedData(JtlCategory $jtlCategory, SwCategory $swCategory, array &$translations, $langIso = null)
     {
         if (is_null($langIso)) {
-            $langIso = LanguageUtil::map(Shopware()->Shop()->getLocale()->getLocale());
+            $langIso = LocaleUtil::extractLanguageIsoFromLocale(Shopware()->Shop()->getLocale()->getLocale());
+        }else{
+            $langIso = LanguageUtil::convert(null, $langIso);
         }
 
         // Attribute
@@ -298,26 +301,36 @@ class Category extends DataMapper
             }
 
             foreach ($jtlAttribute->getI18ns() as $attributeI18n) {
-                if ($attributeI18n->getLanguageISO() === $langIso) {
+                if (LanguageUtil::convert(null, $attributeI18n->getLanguageISO()) === $langIso) {
 
-                    // Active fix
-                    $allowedActiveValues = array('0', '1', 0, 1, false, true);
-                    if (in_array(strtolower($attributeI18n->getName()), [CategoryAttr::IS_ACTIVE, 'isactive'])
-                        && in_array($attributeI18n->getValue(), $allowedActiveValues, true)) {
-                        $swCategory->setActive((bool)$attributeI18n->getValue());
+                    if (CategoryAttr::isSpecialAttribute($attributeI18n->getName())) {
 
-                        continue;
-                    }
+                        // Active fix
+                        $allowedActiveValues = array('0', '1', 0, 1, false, true);
+                        if (in_array(strtolower($attributeI18n->getName()), [CategoryAttr::IS_ACTIVE, 'isactive'])
+                            && in_array($attributeI18n->getValue(), $allowedActiveValues, true)) {
+                            $swCategory->setActive((bool)$attributeI18n->getValue());
+                        }
 
-                    // Cms Headline
-                    if (in_array(strtolower($attributeI18n->getName()), [CategoryAttr::CMS_HEADLINE, 'cmsheadline'])) {
-                        $swCategory->setCmsHeadline($attributeI18n->getValue());
+                        // Cms Headline
+                        if (in_array(strtolower($attributeI18n->getName()),
+                            [CategoryAttr::CMS_HEADLINE, 'cmsheadline'])) {
+                            $swCategory->setCmsHeadline($attributeI18n->getValue());
 
-                        foreach ($jtlAttribute->getI18ns() as $i18n) {
-                            if ($i18n->getLanguageISO() === $langIso) {
-                                continue;
+                            foreach ($jtlAttribute->getI18ns() as $i18n) {
+                                if (LanguageUtil::convert(null, $i18n->getLanguageISO()) === $langIso) {
+                                    continue;
+                                }
+                                $translations[$i18n->getLanguageISO()]['category']['cmsheadline'] = $i18n->getValue();
                             }
-                            $translations[$i18n->getLanguageISO()]['category']['cmsheadline'] = $i18n->getValue();
+                        }
+
+                        if ($attributeI18n->getName() === CategoryAttr::IS_BLOG) {
+                            $swCategory->setBlog((bool)$attributeI18n->getValue());
+                        }
+
+                        if ($attributeI18n->getName() === CategoryAttr::LIMIT_TO_SHOPS) {
+                            $swCategory->setShops($attributeI18n->getValue());
                         }
 
                         continue;
@@ -367,7 +380,7 @@ class Category extends DataMapper
     public function prepareCategoryMapping(JtlCategory $jtlCategory, SwCategory $swCategory)
     {
         foreach ($jtlCategory->getI18ns() as $i18n) {
-            if (strlen($i18n->getLanguageISO()) > 0 && LanguageUtil::map(null, null, $i18n->getLanguageISO()) !== Shopware()->Shop()->getLocale()->getLocale()) {
+            if (strlen($i18n->getLanguageISO()) > 0 && ShopUtil::isShopwareDefaultLanguage($i18n->getLanguageISO()) === false) {
                 $categoryMappingSW = CategoryMappingUtil::findCategoryMappingByParent($swCategory->getId(), $i18n->getLanguageISO());
 
                 if (is_null($categoryMappingSW)) {
@@ -424,20 +437,17 @@ class Category extends DataMapper
         foreach ($jtlCategory->getI18ns() as $i18n) {
             $langIso2B = $i18n->getLanguageISO();
             $langIso1 = LanguageUtil::convert(null, $langIso2B);
-            if ($langIso2B === LanguageUtil::map(ShopUtil::locale()->getLocale())) {
+            if (ShopUtil::isShopwareDefaultLanguage($langIso2B)) {
                 continue;
             }
 
+            $shopMapper = Mmc::getMapper('Shop');
             /** @var \Shopware\Models\Shop\Shop[] $shops */
-            $shops = ShopUtil::entityManager()->getRepository(\Shopware\Models\Shop\Shop::class)->findAll();
-            foreach ($shops as $shop) {
-                if (strpos($shop->getLocale()->getLocale(), $langIso1) !== 0) {
-                    continue;
-                }
+            $shops = $shopMapper->findByLanguageIso($langIso1);
 
+            foreach ($shops as $shop) {
                 $categoryTranslation = array_filter([
                     'description' => $i18n->getName(),
-                    //'cmsheadline' => $i18n->get,
                     'cmstext' => $i18n->getDescription(),
                     'metatitle' => $i18n->getTitleTag(),
                     'metakeywords' => $i18n->getMetaKeywords(),