+++ src/Mapper/Product.php
@@ -34,6 +34,8 @@ use jtl\Connector\Linker\ChecksumLinker;
 use jtl\Connector\Shopware\Mapper\ProductPrice as ProductPriceMapper;
 use jtl\Connector\Shopware\Model\ProductAttr;
 use jtl\Connector\Shopware\Utilities\CategoryMapping as CategoryMappingUtil;
+use Shopware\Models\Customer\PriceGroup;
+use \Shopware\Models\Price\Group as SwGroup;
 use Shopware\Models\Plugin\Plugin;
 use Shopware\Models\Property\Group;
 use Shopware\Models\Property\Option;
@@ -509,7 +511,7 @@ class Product extends DataMapper
 
         // I18n
         foreach ($product->getI18ns() as $i18n) {
-            if ($i18n->getLanguageISO() === LanguageUtil::map(Shopware()->Shop()->getLocale()->getLocale())) {
+            if (ShopUtil::isShopwareDefaultLanguage($i18n->getLanguageISO())) {
                 $productSW->setDescription($i18n->getMetaDescription())
                     ->setDescriptionLong($i18n->getDescription())
                     ->setKeywords($i18n->getMetaKeywords())
@@ -542,7 +544,7 @@ class Product extends DataMapper
                     // Category Mapping
                     if ($useMapping) {
                         foreach ($product->getI18ns() as $i18n) {
-                            if ($i18n->getLanguageISO() !== LanguageUtil::map(Shopware()->Shop()->getLocale()->getLocale()) && strlen($i18n->getName()) > 0) {
+                            if (ShopUtil::isShopwareDefaultLanguage($i18n->getLanguageISO()) === false && strlen($i18n->getName()) > 0) {
                                 $categoryMapping = CategoryMappingUtil::findCategoryMappingByParent($categorySW->getId(), $i18n->getLanguageISO());
                                 if ($categoryMapping !== null) {
                                     $collection->add($categoryMapping);
@@ -724,19 +726,32 @@ class Product extends DataMapper
 
         // Delivery time
         $exists = false;
-        foreach ($product->getI18ns() as $i18n) {
-            if ($i18n->getLanguageISO() === LanguageUtil::map(Shopware()->Shop()->getLocale()->getLocale())) {
-                $days = trim(str_replace(['Tage', 'Days', 'Tag', 'Day'], '', $i18n->getDeliveryStatus()));
-                if (strlen($days) > 0 && $days !== '0') {
-                    $detailSW->setShippingTime($days);
-                    $exists = true;
-                    break;
+        $useHandlingTimeOnly = (bool)Application()->getConfig()->get('product.push.use_handling_time_for_shipping', false);
+        if(!$useHandlingTimeOnly) {
+            foreach ($product->getI18ns() as $i18n) {
+                if ($i18n->getLanguageISO() === LanguageUtil::map(Shopware()->Shop()->getLocale()->getLocale())) {
+                    $deliveryStatus = trim(str_replace(['Tage', 'Days', 'Tag', 'Day'], '', $i18n->getDeliveryStatus()));
+                    if ($deliveryStatus !== '' && $deliveryStatus !== '0') {
+                        $detailSW->setShippingTime($deliveryStatus);
+                        $exists = true;
+                        break;
+                    }
                 }
             }
         }
 
         if (!$exists) {
-            $detailSW->setShippingTime($product->calculateHandlingTime());
+            $shippingTime = $product->calculateHandlingTime();
+            $considerNextAvailableInflowDate = (bool)Application()->getConfig()->get('product.push.consider_supplier_inflow_date_for_shipping', true);
+            if($product->getStockLevel()->getStockLevel() <= 0 && $considerNextAvailableInflowDate && !is_null($product->getNextAvailableInflowDate())) {
+                $inflow = new \DateTime($product->getNextAvailableInflowDate()->format('Y-m-d'));
+                $today = new \DateTime((new \DateTime())->format('Y-m-d'));
+                if($inflow->getTimestamp() - $today->getTimestamp() > 0) {
+                    $shippingTime = $inflow->diff($today)->days;
+                }
+            }
+
+            $detailSW->setShippingTime($shippingTime);
         }
 
         // Last stock
@@ -771,7 +786,7 @@ class Product extends DataMapper
         foreach ($product->getVariations() as $variation) {
             $variationName = null;
             foreach ($variation->getI18ns() as $variationI18n) {
-                if ($variationI18n->getLanguageISO() === LanguageUtil::map(Shopware()->Shop()->getLocale()->getLocale())) {
+                if (ShopUtil::isShopwareDefaultLanguage($variationI18n->getLanguageISO())) {
                     $variationName = $variationI18n->getName();
                 }
             }
@@ -781,7 +796,7 @@ class Product extends DataMapper
                 foreach ($variation->getValues() as $variationValue) {
                     $name = null;
                     foreach ($variationValue->getI18ns() as $variationValueI18n) {
-                        if ($variationValueI18n->getLanguageISO() === LanguageUtil::map(Shopware()->Shop()->getLocale()->getLocale())) {
+                        if (ShopUtil::isShopwareDefaultLanguage($variationValueI18n->getLanguageISO())) {
                             $name = $variationValueI18n->getName();
                         }
                     }
@@ -828,7 +843,7 @@ class Product extends DataMapper
             }
 
             foreach ($attribute->getI18ns() as $attributeI18n) {
-                if ($attributeI18n->getLanguageISO() === LanguageUtil::map(Shopware()->Shop()->getLocale()->getLocale())) {
+                if (ShopUtil::isShopwareDefaultLanguage($attributeI18n->getLanguageISO())) {
                     $lcAttributeName = strtolower($attributeI18n->getName());
                     $attributeValue = $attributeI18n->getValue();
 
@@ -871,6 +886,20 @@ class Product extends DataMapper
                     if (in_array($lcAttributeName, [ProductAttr::PSEUDO_SALES, 'sw_pseudo_sales'])) {
                         $article->setPseudoSales((int)$attributeValue);
 
+                        continue;
+                    }
+
+                    if ($lcAttributeName === ProductAttr::PRICE_GROUP_ID) {
+                        if (empty($attributeValue)) {
+                            $article->setPriceGroupActive(false);
+                        }else{
+                            $article->setPriceGroupActive(true);
+                            $priceGroupId = (int)$attributeValue;
+                            $priceGroupSW = Shopware()->Models()->getRepository(SwGroup::class)->find($priceGroupId);
+                            if ($priceGroupSW instanceof SwGroup) {
+                                $article->setPriceGroup($priceGroupSW);
+                            }
+                        }
                         continue;
                     }
 
@@ -1005,7 +1034,7 @@ class Product extends DataMapper
                 $variationName = null;
                 $variationValueName = null;
                 foreach ($variation->getI18ns() as $variationI18n) {
-                    if ($variationI18n->getLanguageISO() === LanguageUtil::map(Shopware()->Shop()->getLocale()->getLocale())) {
+                    if (ShopUtil::isShopwareDefaultLanguage($variationI18n->getLanguageISO())) {
                         $variationName = $variationI18n->getName();
                     }
                 }
@@ -1025,7 +1054,7 @@ class Product extends DataMapper
 
                 foreach ($variation->getValues() as $i => $variationValue) {
                     foreach ($variationValue->getI18ns() as $variationValueI18n) {
-                        if ($variationValueI18n->getLanguageISO() === LanguageUtil::map(Shopware()->Shop()->getLocale()->getLocale())) {
+                        if (ShopUtil::isShopwareDefaultLanguage($variationValueI18n->getLanguageISO())) {
                             $variationValueName = $variationValueI18n->getName();
                         }
                     }
@@ -1246,21 +1275,19 @@ class Product extends DataMapper
         }
 
         $translationService = ShopUtil::translationService();
-        /** @var \Shopware\Models\Shop\Shop[] $shops */
-        $shops = ShopUtil::entityManager()->getRepository(\Shopware\Models\Shop\Shop::class)->findAll();
 
         foreach ($translations as $langIso2B => $translation) {
             /** @var \Shopware\Models\Shop\Locale $locale */
-            if ($langIso2B === LanguageUtil::map(ShopUtil::locale()->getLocale())) {
+            $langIso1 = LanguageUtil::convert(null, $langIso2B);
+            if ($langIso1 === LocaleUtil::extractLanguageIsoFromLocale(ShopUtil::locale()->getLocale())) {
                 continue;
             }
 
-            $langIso1 = LanguageUtil::convert(null, $langIso2B);
-            foreach ($shops as $shop) {
-                if (strpos($shop->getLocale()->getLocale(), $langIso1) !== 0) {
-                    continue;
-                }
+            /** @var \Shopware\Models\Shop\Shop[] $shops */
+            $shopMapper = Mmc::getMapper('Shop');
+            $shops = $shopMapper->findByLanguageIso($langIso1);
 
+            foreach ($shops as $shop) {
                 if ($merge) {
                     $savedTranslation = $translationService->read($shop->getId(), $type, $key);
                     $translation = array_merge($savedTranslation, $translation);
@@ -1287,7 +1314,7 @@ class Product extends DataMapper
         $data = [];
         foreach ($product->getI18ns() as $i18n) {
             $langIso = $i18n->getLanguageISO();
-            if ($langIso === LanguageUtil::map(ShopUtil::locale()->getLocale())) {
+            if (ShopUtil::isShopwareDefaultLanguage($langIso)) {
                 continue;
             }
 
@@ -1327,7 +1354,7 @@ class Product extends DataMapper
         foreach ($product->getAttributes() as $attribute) {
             foreach ($attribute->getI18ns() as $attrI18n) {
                 $langIso = $attrI18n->getLanguageISO();
-                if ($langIso === LanguageUtil::map(ShopUtil::locale()->getLocale())) {
+                if (ShopUtil::isShopwareDefaultLanguage($langIso)) {
                     continue;
                 }
 
@@ -1351,16 +1378,16 @@ class Product extends DataMapper
             if (!is_null($unitSW)) {
                 foreach ($unitSW->getI18ns() as $unitI18n) {
                     $langIso = $unitI18n->getLanguageIso();
-                    if ($langIso === LanguageUtil::map(ShopUtil::locale()->getLocale())) {
+                    if (ShopUtil::isShopwareDefaultLanguage($langIso)) {
                         continue;
                     }
-                }
 
-                if (!isset($data[$langIso])) {
-                    $data[$langIso] = $this->initVariantTranslation();
-                }
+                    if (!isset($data[$langIso])) {
+                        $data[$langIso] = $this->initVariantTranslation();
+                    }
 
-                $data[$langIso]['packUnit'] = $unitI18n->getName();
+                    $data[$langIso]['packUnit'] = $unitI18n->getName();
+                }
             }
         }
 
@@ -1408,10 +1435,10 @@ class Product extends DataMapper
             // Get default translation values
             $variations = array();
             $values = array();
-            $defaultIso = LanguageUtil::map(Shopware()->Shop()->getLocale()->getLocale());
+
             foreach ($product->getVariations() as $variation) {
                 foreach ($variation->getI18ns() as $variationI18n) {
-                    if ($variationI18n->getLanguageISO() === $defaultIso) {
+                    if (ShopUtil::isShopwareDefaultLanguage($variationI18n->getLanguageISO())) {
                         $variations[$variationI18n->getName()] = $variation->getId()->getHost();
                         break;
                     }
@@ -1419,7 +1446,7 @@ class Product extends DataMapper
 
                 foreach ($variation->getValues() as $value) {
                     foreach ($value->getI18ns() as $valueI18n) {
-                        if ($valueI18n->getLanguageISO() === $defaultIso) {
+                        if (ShopUtil::isShopwareDefaultLanguage($valueI18n->getLanguageISO())) {
                             $values[$variation->getId()->getHost()][$valueI18n->getName()] = $value->getId()->getHost();
                             break;
                         }
@@ -1430,7 +1457,7 @@ class Product extends DataMapper
             // Write non default translation values
             foreach ($product->getVariations() as $variation) {
                 foreach ($variation->getI18ns() as $variationI18n) {
-                    if ($variationI18n->getLanguageISO() !== $defaultIso) {
+                    if (ShopUtil::isShopwareDefaultLanguage($variationI18n->getLanguageISO()) !== false) {
                         foreach ($confiSetSW->getGroups() as $groupSW) {
                             if (isset($variations[$groupSW->getName()]) && $variations[$groupSW->getName()] == $variation->getId()->getHost()) {
                                 try {
@@ -1445,7 +1472,7 @@ class Product extends DataMapper
 
                 foreach ($variation->getValues() as $value) {
                     foreach ($value->getI18ns() as $valueI18n) {
-                        if ($valueI18n->getLanguageISO() !== $defaultIso) {
+                        if (ShopUtil::isShopwareDefaultLanguage($valueI18n->getLanguageISO())) {
                             foreach ($confiSetSW->getOptions() as $optionSW) {
                                 if (isset($values[$variation->getId()->getHost()][$optionSW->getName()])
                                     && $values[$variation->getId()->getHost()][$optionSW->getName()] == $value->getId()->getHost()) {
@@ -1525,7 +1552,7 @@ class Product extends DataMapper
         foreach ($product->getMediaFiles() as $mediaFile) {
             $name = '';
             foreach ($mediaFile->getI18ns() as $i18n) {
-                if ($i18n->getLanguageIso() === LanguageUtil::map(Shopware()->Shop()->getLocale()->getLocale())) {
+                if (ShopUtil::isShopwareDefaultLanguage($i18n->getLanguageIso())) {
                     $name = $i18n->getName();
                 }
             }