+++ src/Model/ProductAttr.php
@@ -21,6 +21,7 @@ class ProductAttr extends ProductAttrModel
     const SEND_NOTIFICATION = 'send_notification';
     const SHIPPING_FREE = 'shipping_free';
     const CUSTOM_PRODUCTS_TEMPLATE = 'custom_products_template';
+    const PRICE_GROUP_ID = 'price_group_id';
 
     protected $fields = array(
         'id' => 'id',
@@ -57,6 +58,7 @@ class ProductAttr extends ProductAttrModel
             self::SHIPPING_FREE,
             'shippingfree',
             self::CUSTOM_PRODUCTS_TEMPLATE,
+            self::PRICE_GROUP_ID
         ];
     }