+++ src/Utilities/Payment.php
@@ -21,7 +21,8 @@ final class Payment
         PaymentTypes::TYPE_BILLSAFE => 'billsafe_invoice',
         PaymentTypes::TYPE_HEIDELPAY => [
             'hgw_iv',
-            'hgw_papg'
+            'hgw_papg',
+            'hgw_ivb2b'
         ]
     );